separate benchmarks by their respective assemblies
