+++ brainscore/benchmarks/README.md
@@ -1,0 +1,3 @@
+# Brain-Score Benchmarks
+
+**File structure**: one file per assembly with all the benchmarks inside.
